make the clean target a bit more resilient

+++ Makefile
@@ -4,12 +4,12 @@ DOCKER_OPTS = -V --force-recreate --always-recreate-deps --attach-dependencies -
 DOCKER_RESTART_OPTS = -V --force-recreate --always-recreate-deps --attach-dependencies
 
 clean:
-	docker-compose -f genstack.yml down
-	docker-compose -f genstack.yml rm -f
+	docker-compose -f genstack.yml down || echo "docker-compose down failed"
+	docker-compose -f genstack.yml rm -f || echo "docker-comose rm failed"
 	docker volume prune -f
 	docker volume ls | fgrep _local_ | awk '{print $2}' | xargs -I {} docker volume rm -f {}
-	sudo rm -rf srv/tower-analytics-backend/local_*_data
-	sudo rm -rf srv/tower-analytics-backend/local_*_data_backups
+	if [[ -d srv/tower-analytics-backend ]]; then sudo rm -rf srv/tower-analytics-backend/local_*_data; fi;
+	if [[ -d srv/tower-analytics-backend ]]; then sudo rm -rf srv/tower-analytics-backend/local_*_data_backups; fi;
 
 stack: clean
 	python3 tool.py
